train_requirements.txt
numpy==1.21.6
pandas==1.3.5
cloudpickle==2.2.0
fastapi==0.85.2
pydantic==1.10.2
uvicorn==0.19.0
tensorflow-gpu==2.10.1
unidecode==1.3.6
tqdm==4.64.1
scikit-learn==1.0.2
